Create Board class and add basic functionality
Created a board class that's meant to be general enough to also work for
chess in the future. It's meant to create two arrays one for the border,
and the other to store the "game state". I added the methods to make
the actual board and then tested to confirm it will work.